Implement CTTZ (Count Trailing Zeros) function using compiler builtins.  This makes switch statements on a Type much faster.
Create AUDITCOMPILER preprocessor name.  When this is set to 1, startup checks to make sure CTTZ works correctly and signed >> properly sign-extends.  This should be run once on each supported architecture and can then be turned off.

Rewrite bp() to do a fast lookup in a table.  It was taking around 1.5% of overall time.  Create AUDITBP preprocessor name to verify that bp() is not called with other than a single-bit mask.  No need to check that any further; it will be removed in future.

Modify selected switch statements on the Type field to apply CTTZ so that the switch can use a branch table efficiently.

There seemed to be testcases for all features where switch statements were changed, except for ": applied to Unicode characters; added these.

Reviewing jtmult(), I noticed that it would incorrectly fail on MAXNEG64 * 1; also that it always did a full 64-bit multiply emulation even though most inputs wouldn't need it.  I rewrote the function & tested it fully.

added #if defined(OBSOLETE)   (which always tests false) to indicate code slated for future removal.