(r=.2.5) ] *&r  crashed from keyboard (untestable in script)
The sequence

9!:3[5
r=.2.5
(r=.2.5) ] *&r

crashed when entered from the keyboard.  The name r was deleted by the nvr-stack processing, but it was referred to in the result; so displaying the result crashed.  Fix is to tpush the nvr names rather than calling fa() directly, letting the name survive till the next EPILOG, by which time all references will have gone away.

This cannot be tested in a script.