Performance improvements, mostly by reducing usecount twiddling
1. jtunquote (execute named entity) did a full recursive increment of usecounts in
the entity.  Now we only increment the usecount in the top level.

2. jt structure is rearranged to cluster byte flags, and to put names that are used
together into the same cache linepair

3. See if a line of an explicit definition cannot possibly contribute to the result;
if so, don't bother protecting its result from tpop

4. Define rat1(), which is like rat() except that it increments usecount only at the
top level.  Verbs use rat() as a signal that their results incorporate some portion
of their inputs, which means that callers cannot reuse any block that contributed
to that result.  Replace rat with rat1 when this is the only reason for using
rat.  So far, box, boxopen, and link are the only places we have made this change, but
others should follow.  Eventually rat1 should be replaced by a different (non-usecount)
mechanism for indicating that a result incorporates some input.

5. [AC]S[12] functions performed an EPILOG function if the operation needed to be
performed on multiple cells.  But F?RANK, which is called by [AC]S[12], already
does EPILOG.  So there is no need for another.  Rather than delete the EPILOG (in
case I am wrong), I replaced it with EPILOGNULL which does nothing.

6. Change MEMAUDIT from an ordinal level to a bitmask

7. Rewrite spfree() to avoid a second pass over the block chain if there are no big
blocks to free, and to reduce the third pass to a pass only over the base blocks

8. For SYMB tables, allocate a length of 1+prime rather than prime, and change
bucket calculation accordingly.  This eliminates the collisions resulting from
grouping buckets 0 and 1 together