Use flags in derived verbs; and a bug fix
As part of the preparation for thoroughgoing inplacing, allow flags to be passed in to (A|C)DERIV.  Then, make these flags correct for VIRS1/VIRS2.  This takes special opcode-checking code out of jtqq and moves it to the special-coded modifiers, which then communicate through verb flags.

Not much of a change, but it does allow *!.f"r to run as fast as *"r.  The main improvement is code simplification, and creation of the pathway that inplacing will need.

Also fixed a bug introduced recently in support of for_xyz.