Add type flag for value-only name (like x, y, u.); remove scaffold code
Add an upper type-flag to NAME types indicating that the name is a value-only name executing in an explicit definition.  Perhaps not worth doing, but it saves a few instructions in the parser path for names, so I did it.

Also remove scaffold code (formerly commented out) used in previous testing.