Start full implementation of inplacing.  This implements fork.
The infrastructure for inplacing is ready.  The job now is to go through the code base, turning on inplacing for every call for which it is valid.  We will probably never get to all of them, but we can be content if we finish the structural modifiers.

The only verb supporting inplacing now is (x , y)