Add 9!:53 to prevent incomplete assignment; default to safe mode
Incomplete assignment occurs when assignment to a zombie modifies the zombiename inplace but does not complete the assignment.  Example is

a =: a (something@:error@:,) 'a'

which appends to name a before executing error, and thus never runs something.  This could cause confusion for greenhorns, so inplacing zombienames is disabled for compounds unless enabled by 9!:53 (1).