Introduce RESTRICTI
We use RESTRICT for applications where buffers are truly separate, and also for in-place operations such as *z++ = *x++ + *y++ where z=y in the hope that the compiler will not generate incorrect code - a hope that is fulfilled when there are plenty of registers, but not on 8-register 32-bit architectures.

This change creates the keyword RESTRICTI which is used when the operation may be in-place, and reserves RESTRICT for truly separate buffers.  RESTRICT is then honored on all architectures, and only RESTRICTI need be disabled on small architectures.