Avoid extra fetch for tolerant comparison
Now, rather than fetching the possible neighbor, we calculate the 2 neighbors (as in 8.05) and then figure out the address of a possible neighbor.  We then fetch the possible neighbor and process it.  It may be the same as the base bucket, but it that case processing it will be very quick.