Fix memory leak in ]^:0:
The error was related to inplacing.  When an in-place argument was sent into gc(), which protects it from being freed until the
current level of execution completes, there is code to detect when the argument was far up the tpop stack.  In that case, the
code simply decremented the usecount and restored the inplaceability rather than putting the block on the tpop stack.  This
worked for the block, but if it had descendants, the usecount of the descendants was not decremented, and the memory was lost.

The fix is to traverse the argument to decrement the usecount in each descendant.  This simple approach fails, though, because some paths,
including sparse amend and rational matrix operations, fiddle with some of the descendants, leaving them younger than the enclosing
argument, with the result that when usecount is decremented, the component is freed while still being pointed to.

The final solution is, in effect, a hybrid of the old method of putting all descendants on the stack and the new one of adjusting usecount.  We
provisionally decrement the usecount, but if it would go to 0, we don't delete the block, but instead push it onto the tpop stack for eventual free.

This should eventually turn into a recursive usecount strategy wherein a block can be marked so as to represent a decrement of all containing blocks.