Reserve RZ for error detection; add FINDNULLRET compile option
RZ was promiscuously used for error returns and for general returns of the value 0.  I separated these, making all the general returns into if-tests.

The option FINDNULLRET in j.h, when set, causes RZ to program-check if the argument is 0 and no error has been set.  This can be used to find errors that produce no displayed result.