Fix bug in #"2 2 etc; expand scope of jtrank2ex
Dyad # $ {. }. were not sophisticated enough in their support of IRS, in that the internal left rank was ignored.  The solution chosen here is to give jtrank2ex two rank-pairs, so that it can produce up to 4 nested loops.  This has less overhead than nesting calls to rank2ex, and should improve performance for u"n1"n2 which nw does not have to call jtrank2ex for the nested call to jtrank2.

Future work may further streamline u"n1"n2 to avoid the nested call to jtrank2 altogether.