Fix -&>/~ ((<"_1) ,._1 1) error; clear warnings
1.  There was an inplacing problem, in that if looking inside a boxed noun returned an inplaceable contents,
that contents was inplaced even if the enclosing box was flagged as not inplaceable.  Such contents occurred only
as the result of <"r yy.  The fix is to make > turn off inplacing if it opens a single box.  The same code is used for
{:: which is the other place the problem could occur.  Eventually we should have > support inplacing so that
the contents can be left inplaceable if that will not cause error.

2. viavx.c uses the name tltr.  Initializing this without warnings is tricky because different platforms need different initialization.  Here is another attempt.