Fix 2!:2 related reference count error
The 2!:2 foreign conjunction is currently broken on Linux.
The jthostio function calls add2 which creates a literal
string, copies the command line into it, and updates
the internal list of open files to contain the file handle
and command line.

The existing code does not increment the reference count on
the string literal. When it is later unreferenced in add2
memory instability results. This effects can be seen as random
segfaults, parser errors and 1!:20 crashing after 2!:2 is used.

The fix here is to bump the reference count so it is correctly
handled when the reverse is done at the end of the function.

Although this fixes 2!:2 and 1!:20 the 1!:1 read call will still
fail on the returned file handle due to issues there where it
attempts to seek on the file handle, but pipes don't support
seeking. The handles can be used from FFI calls to fread however.