Reboot project to overhaul the make2 build system.
The goal is to provide a build process that is friendly to package
maintainers for arbitrary Linux distributions, while maintaining the
current status quo for non-Linux targets.

In particular, the project aims to improved the following perceived
deficiences of current make2:

1. Ad-hoc, idiosyncratic shell scripts;
2. Fragile custom code for compiler detection;
3. Ignores host's CFLAGS, LDFLAGS etc;
4. Non-obvious availability of other build flags, e.g. `USE_OPENMP` etc;
5. Non-error exit code on failure;
6. Unsavory reputation amongst existing package maintainers.

The above issues are adressable by updates to the build scripts alone.
However, there are further issues that require supporting changes
elsewhere:

7. Build non-determinism (due to `__DATE__` and `__TIME__` macro usage);
8. Hostile to FHS-conforming installs.

Importantly, in the course of working on the above, we impose some
important constraints on ourself:

1. Build system changes must not introduce unintended changes in build
   products;
2. The current, officially supported build environments must not be
   broken;
3. Changes should be as conservative as possible to achieve goals.

This commit introduces a temporary meta-build tool, `jbuild.scm`, to
help maintain the above mentioned constraints. See that file for
details.