jbuild.scm: Instrument in automatic build verification
Executing jbuild.scm as a script will build J from source, hash the
output, and display the result, with a comparison against a known good
target hash.

Note that, unfortunately, the diff in this commit got mixed with a large
re-indentation fix, but the bulk of semantic changes take place below
the `;;; Execution' section.