make3: Explicitly define guix channel revision
For 100% build reproducibility we also need to ensure that we build
against a stable set of dependencies. This commit defines an explicit
guix channel revision against which the build script runs.

In particular, without an explicit channel revision and when run on
different machines, the build script is likely to produce different
outputs, causing a mismatch with the target hash. In fact, the same may
happen on a single machine, if invoked across `guix pull` updates.