gmp emergency memory pool with wsfull errors for libgmp
first draft -- this gets us wsfull errors

no test suite yet

may be overly paranoid in some contexts and not paranoid enough in other contexts

(macros which return I or D values do not have a supporting convention for representing error values. So wsfull conditions during those macros are ignored here.)

A corresponding issue is that "exponential functions" like ! and ^ might generate results too large for the pool to cushion. These will need guard statements which throw an error early based on an estimate of the size of the result. This commit defines GMPMAXSZ for that purpose.