🔀 Do not determine compiler name from file name
The build scripts determine currently the compiler's name from the file
name.  The default compiler name tried is cc and if this is found,
through the use of readlink, the real file name is determined and
tested.

The problem is if cc does not point to the compiler driver.  This is the
case if ccache is used which has cc installed as a symlink to the ccache
binary.  In that case the current tests assume clang is used which will
fail in case the real cc is actually a symlink to gcc.

The patch fixes this by asking the compiler to identify itself.  Since
the compiler has to be executable for the build process and only gcc and
clang are supported and because these two compilers support the -v flag,
the patch grabs the name of the compiler from that output.  The compiler
variable is repurposed for this.

Tested on Fedora 43.