🐛 Allow prefixes to compiler names
When the -v option is used both gcc and clang by default respond with
a string starting with wither gcc or clang.  Some distributions seem
to modify this behavior, e.g., Ubuntu which prepends "Ubuntu " to
the string.  Account for these variations when determining which
compiler is used in the build scripts.