🧱 Remove warnings gcc does not know
The build scripts enable some warnings gcc does not know.  Normally
these are ignored because of a special feature in gcc which does not
complain about unknown warnings if they start with -Wno-.  But if
an error is found during the compilation the warnings are listed as
unknown which might confuse users.  Just avoid the warnings.